refactor(server): make execution service runtime-only
