+++ crates/harness-server/src/router/tests.rs
@@ -52,17 +52,7 @@ async fn make_test_state_with_config_and_registry(
         crate::services::project::DefaultProjectService::new(_project_svc_tmp, dir.to_path_buf());
     let task_svc = crate::services::task::DefaultTaskService::new(tasks.clone());
     let execution_svc = crate::services::execution::DefaultExecutionService::new(
-        tasks.clone(),
-        server.agent_registry.clone(),
         Arc::new(server.config.clone()),
-        Default::default(),
-        events.clone(),
-        vec![],
-        None,
-        Arc::new(crate::task_queue::TaskQueue::new(&Default::default())),
-        Arc::new(crate::task_queue::TaskQueue::new(&Default::default())),
-        None,
-        None,
         None,
         None,
         vec![],
